refactor: split views.py
